Develop

+++ src/main/java/com/esgi/TestController.java
@@ -9,49 +9,64 @@ import java.nio.charset.StandardCharsets;
 
 @RestController
 public class TestController {
-    @GetMapping("/test")
-    public ResponseEntity<String> test() throws IOException {
-        String pythonScript = "print('Hello')\nprint('World')";
+
+    @GetMapping("/c-test")
+    public final ResponseEntity<String> cTest() {
+        final String cScript = """
+                #include <stdio.h>
+                int main() {
+                   printf("Hello, cWorld!");
+                   return 0;
+                }
+                """;
         final long startTime = System.nanoTime();
 
+        createFile("/home/ec2-user/c_src/app.c", cScript);
+        final String output = executeCommand("/home/ec2-user/run_c.sh");
+
+        final long endTime = System.nanoTime();
+
+        final long duration = (endTime - startTime) / 1_000_000;
+        return ResponseEntity.ok(output + " " + duration + "ms");
+    }
+
+    @GetMapping("/python-test")
+    public final ResponseEntity<String> pythonTest() {
+        final String pythonScript = "print('Hello')\nprint('World')";
+        final long startTime = System.nanoTime();
+
+        createFile("/home/ec2-user/python_src/app.py", pythonScript);
+        final String output = executeCommand("/home/ec2-user/run_python.sh");
+
+        final long endTime = System.nanoTime();
+
+        final long duration = (endTime - startTime) / 1_000_000;
+        return ResponseEntity.ok(output + " " + duration + "ms");
+    }
+
+    private void createFile(String filePath, String content) {
+        try {
+            File targetFile = new File(filePath);
+            if (targetFile.exists()) {
+                targetFile.delete();
+                targetFile.createNewFile();
+            }
+            FileOutputStream outputStream = new FileOutputStream(targetFile);
+            outputStream.write(content.getBytes(StandardCharsets.UTF_8));
+            outputStream.close();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+
+    }
+
+    private String executeCommand(String command) {
         ProcessBuilder processBuilder = new ProcessBuilder();
+        processBuilder.command(command);
 
-        final String FILE_NAME = "/home/ec2-user/python_src/test.py";
-        File targetFile = new File(FILE_NAME);
-        targetFile.delete();
-        targetFile.createNewFile();
-
-        FileOutputStream outputStream = new FileOutputStream(targetFile);
-        outputStream.write(pythonScript.getBytes(StandardCharsets.UTF_8));
-        outputStream.close();
-
-
-        // -- Linux --
-
-        // Run a shell command
-//        processBuilder.command("bash", "-c", "ls /home/mkyong/");
-
-//        try {
-//            new ProcessBuilder().command("echo \"" + pythonScript + "\" > /home/ec2-user/python_src/test.py").start();
-//        } catch (IOException e) {
-//            e.printStackTrace();
-//        }
-        // Run a shell script
-        processBuilder.command("/home/ec2-user/run_python.sh");
-
-        // -- Windows --
-
-        // Run a command
-//        processBuilder.command("cmd.exe", "/c", "dir C:\\Users\\ademr");
-
-        // Run a bat file
-        //processBuilder.command("C:\\Users\\mkyong\\hello.bat");
         StringBuilder output = new StringBuilder();
         try {
-
             Process process = processBuilder.start();
-
-
             BufferedReader reader = new BufferedReader(
                     new InputStreamReader(process.getInputStream()));
 
@@ -62,15 +77,12 @@ public class TestController {
 
             int exitVal = process.waitFor();
             if (exitVal != 0) {
-                return ResponseEntity.internalServerError().body("Cannot run python script.");
+                throw new UnsupportedOperationException("Cannot run script");
             }
 
         } catch (IOException | InterruptedException e) {
             e.printStackTrace();
         }
-        long endTime = System.nanoTime();
-
-        long duration = (endTime - startTime) / 1000000;  //divide by 1000000 to get milliseconds.
-        return ResponseEntity.ok(output + " " + duration + "ms");
+        return output.toString();
     }
 }
